+++ src/ElementMap.php
@@ -40,7 +40,7 @@ class ElementMap extends Plugin
 	public function renderEntryElementMap(array &$context)
 	{
 		if ($context['entry']['id'] != null) {
-			return $this->renderer->render($context['entry']['id']);
+			return $this->renderer->render($context['entry']['id'], $context['site']['id']);
 		}
 	}
 
@@ -51,7 +51,7 @@ class ElementMap extends Plugin
 	public function renderCategoryElementMap(array &$context)
 	{
 		if ($context['category']['id'] != null) {
-			return $this->renderer->render($context['category']['id']);
+			return $this->renderer->render($context['category']['id'], $context['site']['id']);
 		}
 	}
 
@@ -62,7 +62,7 @@ class ElementMap extends Plugin
 	public function renderUserElementMap(array &$context)
 	{
 		if ($context['user']['id'] != null) {
-			return $this->renderer->render($context['user']['id']);
+			return $this->renderer->render($context['user']['id'], $context['site']['id']);
 		}
 	}
 
@@ -73,7 +73,7 @@ class ElementMap extends Plugin
 	public function renderProductElementMap(array &$context)
 	{
 		if ($context['product']['id'] != null) {
-			return $this->renderer->render($context['product']['id']);
+			return $this->renderer->render($context['product']['id'], $context['site']['id']);
 		}
 	}
 }